+++ database/migrations/2020_02_21_164640_create_events_table.php
@@ -17,7 +17,7 @@ class CreateEventsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('module_instance_id');
             $table->text('event');
-            $table->json('data');
+            $table->json('data')->nullable();
             $table->timestamps();
         });
     }