+++ tests/Http/Controllers/Auth/RegisterControllerTest.php
@@ -23,7 +23,12 @@ class RegisterControllerTest extends TestCase
         $this->assertDatabaseHas('control_data_user', [
             'first_name' => 'TestName1',
             'last_name' => 'TestName2',
+            'preferred_name' => 'TestName1 TestName2',
             'email' => 'email@email.com'
+        ]);
+
+        $this->assertDatabaseHas('control_users', [
+            'data_provider_id' => DataUser::where('first_name', 'TestName1')->firstOrFail()->id()
         ]);
 
         $this->assertDatabaseHas('users', [
@@ -31,7 +36,78 @@ class RegisterControllerTest extends TestCase
                 DataUser::where('first_name', 'TestName1')->firstOrFail()->id()
             )->firstOrFail()->id()
         ]);
-
     }
+
+    /** @test */
+    public function an_email_is_required(){
+        $response = $this->post('/register', [
+            'first_name' => 'TestName1',
+            'last_name' => 'TestName2',
+            'password' => 'secret',
+            'password_confirmation' => 'secret',
+        ]);
+        $response->assertSessionHasErrors(['email']);
+    }
+
+    /** @test */
+    public function a_first_name_is_required(){
+        $response = $this->post('/register', [
+            'email' => 'test@test.com',
+            'last_name' => 'TestName2',
+            'password' => 'secret',
+            'password_confirmation' => 'secret',
+        ]);
+        $response->assertSessionHasErrors(['first_name']);
+    }
+
+    /** @test */
+    public function a_last_name_is_required(){
+        $response = $this->post('/register', [
+            'email' => 'test@test.com',
+            'first_name' => 'TestName2',
+            'password' => 'secret',
+            'password_confirmation' => 'secret',
+        ]);
+        $response->assertSessionHasErrors(['last_name']);
+    }
+
+    /** @test */
+    public function a_password_is_required(){
+        $response = $this->post('/register', [
+            'first_name' => 'TestName1',
+            'last_name' => 'TestName2',
+            'email' => 'test@test.com',
+        ]);
+        $response->assertSessionHasErrors(['password']);
+    }
+
+    /** @test */
+    public function password_must_be_confirmed(){
+        $response = $this->post('/register', [
+            'first_name' => 'TestName1',
+            'last_name' => 'TestName2',
+            'email' => 'test@test.com',
+            'password' => 'abc123',
+            'password_confirmation' => '123abc'
+        ]);
+        $response->assertSessionHasErrors(['password']);
+    }
+
+    /** @test */
+    public function the_email_must_be_unique(){
+        $dataUser = factory(DataUser::class)->create(['email' => 'email@email.com']);
+
+        $response = $this->post('/register', [
+            'first_name' => 'TestName1',
+            'last_name' => 'TestName2',
+            'email' => 'email@email.com',
+            'password' => 'secret',
+            'password_confirmation' => 'secret',
+        ]);
+        $response->assertSessionHasErrors('email');
+        $this->assertEquals(1, DataUser::all()->count());
+    }
+
+
 
 }