+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Completion condition testing
 - See all fired events
 
+### Changed
+- Authentication now returns a group and role when needed by the activity
+- Creating a module instance accepts an 'activity_for' type, user, group or role
+
 ## [1.0.3] - (12/02/2020)
 
 ### Changed